Move hyperopt_conf.py into user_data/
